+++ src/Concerns/InteractsWithIO.php
@@ -2,18 +2,19 @@
 
 namespace Henzeb\Console\Concerns;
 
-use RuntimeException;
-use Illuminate\Support\Facades\App;
-use Illuminate\Console\View\Components\Factory;
 use Illuminate\Console\Concerns\InteractsWithIO as IlluminateInteractsWithIO;
+use Illuminate\Console\View\Components\Factory;
+use Illuminate\Support\Facades\App;
+use RuntimeException;
 
 trait InteractsWithIO
 {
-    use IlluminateInteractsWithIO;
+    use IlluminateInteractsWithIO,
+        HandlesVerbosityOutput;
 
     public function components(): Factory
     {
-        if(version_compare(App::version(), '9.21.0', '>=')) {
+        if (version_compare(App::version(), '9.21.0', '>=')) {
             return resolve(Factory::class, ['output' => $this->getOutput()]);
         }