+++ tests/Unit/Console/Providers/ConsoleServiceProviderTest.php
@@ -2,10 +2,12 @@
 
 namespace Henzeb\Console\Tests\Unit\Console\Providers;
 
-use Orchestra\Testbench\TestCase;
 use Henzeb\Console\Facades\Console;
-use Illuminate\Support\Facades\Artisan;
+use Henzeb\Console\Output\ConsoleOutput;
 use Henzeb\Console\Providers\ConsoleServiceProvider;
+use Illuminate\Contracts\Console\Kernel;
+use Illuminate\Support\Facades\Artisan;
+use Orchestra\Testbench\TestCase;
 
 class ConsoleServiceProviderTest extends TestCase
 {
@@ -17,6 +19,10 @@ class ConsoleServiceProviderTest extends TestCase
     public function testShouldIsolateConsoleWhenUsingArtisanCall()
     {
         $unit = $this;
+        if (method_exists($this->app[Kernel::class], 'rerouteSymfonyCommandEvents')) {
+            $this->app[Kernel::class]->rerouteSymfonyCommandEvents();
+        }
+
         Artisan::command(
             'myApplication',
             function () use ($unit) {
@@ -36,5 +42,29 @@ class ConsoleServiceProviderTest extends TestCase
         );
 
         Artisan::call('myApplication');
+    }
+
+    public function providesFormatterStyleTestCases(): array
+    {
+        return [
+            ['henzeb.console.verbose', "\e[36mtest\e[39m"],
+            ['henzeb.console.very.verbose', "\e[33mtest\e[39m"],
+            ['henzeb.console.debug', "\e[35mtest\e[39m"]
+        ];
+    }
+
+    /**
+     * @return void
+     * @dataProvider providesFormatterStyleTestCases
+     */
+    public function testShouldHaveStyleRegistered(
+        string $tag,
+        string $expected
+    ): void
+    {
+        $output = new ConsoleOutput();
+        $output->getOutput()->setDecorated(true);
+        $formatter = $output->getOutput()->getFormatter();
+        $this->assertEquals($expected, $formatter->format(sprintf('<%s>test</>', $tag)));
     }
 }